Refine Anime.js loader
