+++ hero-animations.js
@@ -1,58 +1,18 @@
-// Import Anime.js from the local node_modules directory so the script works
-// when served by a simple HTTP server without a bundler.
+import { getAnime } from './anime-loader.js';
 let animeModule;
 
 export async function typeSubtitle(element, text, interval = 50) {
   element.textContent = '';
   element.classList.add('typing');
   for (const char of text) {
-    // eslint-disable-next-line no-await-in-loop
     await new Promise((r) => setTimeout(r, interval));
     element.textContent += char;
   }
   element.classList.remove('typing');
 }
 
-const CDN_URL =
-  'https://cdn.jsdelivr.net/npm/animejs@4.0.2/lib/anime.esm.min.js';
-// sha256 hash of the file at CDN_URL (hex encoded)
-const EXPECTED_CDN_HASH =
-  'e8eb5b27f49049d82da9cebd01d3809ea5141f6d524f2960824345e0ca45f237';
-
 async function loadAnime() {
-  try {
-    const res = await fetch('./node_modules/animejs/lib/anime.esm.js', { method: 'HEAD' });
-    if (res.ok) {
-      animeModule = await import('animejs');
-      return;
-    }
-    throw new Error('local file missing');
-  } catch (err) {
-    console.warn('Local Anime.js not found, loading from CDN...', err);
-    try {
-      const res = await fetch(CDN_URL);
-      const text = await res.text();
-      const encoder = new TextEncoder();
-      const data = encoder.encode(text);
-      const hashBuffer = await crypto.subtle.digest('SHA-256', data);
-      const hashArray = Array.from(new Uint8Array(hashBuffer));
-      const hashHex = hashArray
-        .map((b) => b.toString(16).padStart(2, '0'))
-        .join('');
-      if (hashHex !== EXPECTED_CDN_HASH) {
-        console.warn('Anime.js CDN hash mismatch. Animations disabled.');
-        animeModule = null;
-        return;
-      }
-      const blob = new Blob([text], { type: 'application/javascript' });
-      const blobURL = URL.createObjectURL(blob);
-      animeModule = await import(blobURL);
-      URL.revokeObjectURL(blobURL);
-    } catch (cdnErr) {
-      console.error('Failed to load Anime.js from CDN', cdnErr);
-      animeModule = null;
-    }
-  }
+  animeModule = await getAnime();
 }
 
 export async function initHeroAnimations() {
